Krannys-Kitchen-Backend/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-phonenumber-field==5.0.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
gunicorn==20.1.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
phonenumbers==8.12.20
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
twilio==6.55.0
urllib3==1.26.4
whitenoise==5.2.0
